pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pillow-avif-plugin>=1.5.2",
    "pillow>=11.3.0",
    "werkzeug>=3.1.3",
]
